+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Tricky number verifier".
 - Tests for "Crop-circles".
 - Tests for "Nicholas Breakspeare and Hugh of Evesham".
+- Tests for "Nature of triangles".
 
 ### Changed
 - Renaming "Linear Bézier curves" to "Cubic Bézier curves".